+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- When an optional node (for negation without existence) was not at the end of
+  the query, `find` queries could give an empty output (#267).
+
 ## [2.4.6] - 2023-07-26
 
 ### Changed